test/e2e: Refactor CreateAndWaitForNamespace
Move this into the provisioner package (which the e2e code depends on),
so that it can be used in the e2e code and provisioner.
