+++ src/cloud-api-adaptor/test/e2e/assessment_helpers.go
@@ -22,13 +22,10 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	"sigs.k8s.io/e2e-framework/klient"
-	"sigs.k8s.io/e2e-framework/klient/k8s"
 	"sigs.k8s.io/e2e-framework/klient/wait"
 	"sigs.k8s.io/e2e-framework/klient/wait/conditions"
 	"sigs.k8s.io/e2e-framework/pkg/env"
 )
-
-const WAIT_NAMESPACE_AVAILABLE_TIMEOUT = time.Second * 120
 
 func reverseSlice(slice []string) []string {
 	length := len(slice)
@@ -663,58 +660,6 @@ func DeletePod(ctx context.Context, client klient.Client, pod *v1.Pod, tcDelDura
 		return err
 	}
 	return nil
-}
-
-func CreateAndWaitForNamespace(ctx context.Context, client klient.Client, namespaceName string) error {
-	log.Infof("Creating namespace '%s'...", namespaceName)
-	nsObj := v1.Namespace{}
-	nsObj.Name = namespaceName
-	if err := client.Resources().Create(ctx, &nsObj); err != nil {
-		return err
-	}
-
-	if err := waitForNamespaceToBeUseable(ctx, client, namespaceName); err != nil {
-		return err
-	}
-	return nil
-}
-
-func waitForNamespaceToBeUseable(ctx context.Context, client klient.Client, namespaceName string) error {
-	log.Infof("Wait for namespace '%s' be ready...", namespaceName)
-	nsObj := v1.Namespace{}
-	nsObj.Name = namespaceName
-	if err := wait.For(conditions.New(client.Resources()).ResourceMatch(&nsObj, func(object k8s.Object) bool {
-		ns, ok := object.(*v1.Namespace)
-		if !ok {
-			log.Printf("Not a namespace object: %v", object)
-			return false
-		}
-		return ns.Status.Phase == v1.NamespaceActive
-	}), wait.WithTimeout(WAIT_NAMESPACE_AVAILABLE_TIMEOUT)); err != nil {
-		return err
-	}
-
-	// SH: There is a race condition where the default service account isn't ready when we
-	// try and use it #1657, so we want to ensure that it is available before continuing.
-	// As the serviceAccount doesn't have a status I can't seem to use the wait condition to
-	// detect if it is ready, so do things the old-fashioned way
-	log.Infof("Wait for default serviceaccount in namespace '%s'...", namespaceName)
-	var saList v1.ServiceAccountList
-	for start := time.Now(); time.Since(start) < WAIT_NAMESPACE_AVAILABLE_TIMEOUT; {
-		if err := client.Resources(namespaceName).List(ctx, &saList); err != nil {
-			return err
-		}
-		for _, sa := range saList.Items {
-			if sa.ObjectMeta.Name == "default" {
-
-				log.Infof("default serviceAccount exists, namespace '%s' is ready for use", namespaceName)
-				return nil
-			}
-		}
-		log.Tracef("default serviceAccount not found after %.0f seconds", time.Since(start).Seconds())
-		time.Sleep(5 * time.Second)
-	}
-	return fmt.Errorf("default service account not found in namespace '%s' after %.0f seconds wait", namespaceName, WAIT_NAMESPACE_AVAILABLE_TIMEOUT.Seconds())
 }
 
 func DeleteAndWaitForNamespace(ctx context.Context, client klient.Client, namespaceName string) error {
